chore: cleaning file structure for the breadcrumb component
