--- packages/react/src/components/Breadcrumb/Breadcrumb.styles.ts
@@ -1,12 +1,0 @@
-import { styled } from '@project44-manifest/react-styles';
-
-export const StyledBreadcrumb = styled('nav', {
-  overflowX: 'auto',
-});
-
-export const StyledBreadcrumbList = styled('ol', {
-  d: 'flex',
-  listStyle: 'none',
-  margin: 0,
-  p: 0,
-});
