--- packages/react/src/components/Breadcrumb/Breadcrumb.types.ts
@@ -1,10 +1,0 @@
-import type { CSS } from '@project44-manifest/react-styles';
-
-export type BreadcrumbElement = 'nav';
-
-export interface BreadcrumbProps {
-  /**
-   * Theme aware style object
-   */
-  css?: CSS;
-}
