--- packages/react/src/components/Breadcrumb/index.ts
@@ -1,2 +1,0 @@
-export { Breadcrumb } from './Breadcrumb';
-export type { BreadcrumbElement, BreadcrumbProps } from './Breadcrumb.types';
