--- packages/react/src/components/BreadcrumbItem/BreadcrumbItem.types.ts
@@ -1,15 +1,0 @@
-import type { CSS } from '@project44-manifest/react-styles';
-
-export type BreadcrumbItemElement = 'li';
-
-export interface BreadcrumbItemProps {
-  /**
-   * Theme aware style object
-   */
-  css?: CSS;
-
-  /**
-   * Breadcrumb item link href
-   */
-  href: string;
-}
