--- packages/react/src/components/BreadcrumbItem/index.ts
@@ -1,2 +1,0 @@
-export { BreadcrumbItem } from './BreadcrumbItem';
-export type { BreadcrumbItemElement, BreadcrumbItemProps } from './BreadcrumbItem.types';
