+++ packages/react/src/components/breadcrumb/Breadcrumb.styles.ts
@@ -1,6 +1,17 @@
 import { styled } from '@project44-manifest/react-styles';
 import { Link } from '../Link';
 
+export const StyledBreadcrumb = styled('nav', {
+  overflowX: 'auto',
+});
+
+export const StyledBreadcrumbList = styled('ol', {
+  d: 'flex',
+  listStyle: 'none',
+  margin: 0,
+  p: 0,
+});
+
 export const StyledBreadcrumbItem = styled('li', {
   d: 'flex',
   alignItems: 'center',
